alcove_file: call per file
