+++ c_src/sys/alcove_file_constants.h
@@ -12,7 +12,7 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-const alcove_define_t alcove_file_constants[] = {
+static const alcove_define_t alcove_file_constants[] = {
 #ifdef O_ACCMODE
     ALCOVE_DEFINE(O_ACCMODE),
 #endif
